requirements.txt
PyQt6==6.8.0
pillow==11.1.0
easyocr==1.7.2
python-docx==1.0.0
reportlab==4.0.9
opencv-python==4.8.0.76
